move AzureOpenAiSecretSettings to azureopenai pkg
